# sui-framework: use test instead of build script