--- a/crates/sui-framework/Cargo.toml
+++ b/crates/sui-framework/Cargo.toml
@@ -18,13 +18,13 @@ sui-types.workspace = true
 move-binary-format.workspace = true
 move-core-types.workspace = true
 
-[build-dependencies]
+[dev-dependencies]
 anyhow.workspace = true
 bcs.workspace = true
 regex.workspace = true
-
-sui-move-build.workspace = true
+tempfile.workspace = true
 
 move-binary-format.workspace = true
 move-compiler.workspace = true
 move-package.workspace = true
+sui-move-build.workspace = true