sui-framework: use test instead of build script

    One area that contributes very negatively to our compilation times is
    sui-framework crate's build script that builds the system packages. The
    build script results in duplicate compilations of a number of crates,
    both first and third-party, as well as the actual build time for system
    packages. This patch removes the build script in favor of doing the
    compilation step in a test function.

    Instead of always building the system packages they'll be checked-in to
    the repo resulting in no additional build step. Instead a test function
    will do the compilation and do a comparison against the checked-in
    compiled packages (and docs) vs the newly generated ones, failing if
    they differ. The checked-in packages and docs can trivially be updated
    by running the test with the `UPDATE` envvar set, e.g.
    `UPDATE=1 cargo test -p sui-framework --test build-system-packages`

    Here are some simple timing measurements for running `cargo check` at
    the root of the repo on my laptop with and with-out this patch:

      with-patch:
        clean build: 3m 15s
        changing sui-types: 17.33s
      without-patch:
        clean build:  4m 00s
        changing sui-types:  1m 42s


## Description 

Describe the changes or additions included in this PR.

## Test plan 

How did you test the new or updated feature?

---

## Release notes

Check each box that your changes affect. If none of the boxes relate to your changes, release notes aren't required.

For each box you select, include information after the relevant heading that describes the impact of your changes that a user might notice and any actions they must take to implement updates. 

- [ ] Protocol: 
- [ ] Nodes (Validators and Full nodes): 
- [ ] Indexer: 
- [ ] JSON-RPC: 
- [ ] GraphQL: 
- [ ] CLI: 
- [ ] Rust SDK: